:building_construction: Set up scaffolding of configuration/data + callback flow
This roughly defines the top-level API and supported props for callers.

The caller feeds a form configuration and optionally an existing
submission with declarative form field values. Configuration and values
need to be processed to evaluate any component-level logic, and after
evaluation, the form field tree can finally be rendered.

We intend to provide a callback to the caller for when a form field
value changes, providing a full view of all the data in the form so
that the caller can kick off whatever is needed as a reaction.

Form state management will be done with Formik, so we can defer the
onSubmit callback and let it be passed in as a prop, which will only
trigger when all client-side validation passes.

Challenges to solve at some point:

1. editgrid component, probably the change context will need to include
   the parent component as well. We also need to check what the logic
   behaviour is, if it can look outside its scope.
2. Currently we only support simple show/hide client-side logic, but
   there are more complex situations possible we might want to use in
   the future. The logic evaluation should ideally be single-pass and
   guaranteed to resolve (cycle detection needed).
3. Styling/CSS and the default component library for the actual
form field rendering.
4. We need a generic way to pass our onChange callback into
   Formik's change handlers at the field level. This is usually
   easier to follow and more performant compared to useEffect
   tricks. Perhaps we need our own Field component build
   around useField.