:sparkles: Port over TextField component and utilities from open-forms-sdk
Adapted for TypeScript and slightly different approach in how
the Formik hooks/components are used, with lessons learned
in the SDK.

This results in a succesful build, but untested if it also
works at runtime. For that, first upgrade Storybook to
v7 so we have a platform to render/test these components.