:technologist: Deal with circular imports
The nature of our component tree requires recursion, but importing the
generic render component in specific files causes circular import
errors that webpack can't handle.

To mitigate this, a plugin is added to detect and prevent circular
imports, and the code is restructured to pass the generic component
itself as a prop instead for rendering child components.

This may get reworked in the future, but it may also not, depending on
how (technical) requirements evolve.