:sparkles: Include SCSS overrides for validation errors from SDK
This pulls in microscope-sass as a (temporary) dependency for parity
with the SDK and making it easier to apply strict BEM selectors
without having to manually implement this again. The module is provided
internally and for now forwards the library module and will at some
point be extracted into a standalone library again.

The validation error styles from the SDK are copied over.