:technologist: Set up tooling to render custom/reference stories in the same way
Certain behaviours are expected to match the reference implementation.

We can define stories in a way that one is renderer with our
implementation, and another is rendered with the reference Formio.js
SDK, with the versions used in the Open Forms SDK. This way, we
guarantee that our (interaction) tests are written correctly by
validating the Formio behaviour, and these same tests are then used
to validate the behaviour of our own implementation.