:sparkles: Initial validation pass implementation
This proof of concept implements email validation options,
except for the 'plugins' validation. We now support the
email format and required/optional validation aspect in
the email component, but most importantly, a pattern
is established for each component type to implement its
validation schema from the available options.

This takes a different approach from Formio itself where
validators are considered global and decoupled from
component types, but that leads to awkwardness. The
implementation details in this library should match the
same functional behaviour compared with formio.js.

For now, it's opted to run validation on blur events,
always, rather than on change events. Formio allows
configuring this on a per-component basis, and I'm
not sure if we can achieve that with Formik + zod.

There's also still the limitation that Formik + zod
only support a single validation error message, while
multiple validations can run at the same time. It's
unclear how important this requirement is at this
stage, though the backend can provide multiple
validation errors in one go that need to be
rendered!