:sparkles: Support conditionally displaying an asterisk for required fields
Uses react-intl for the interpolation when we need to
provide a suffix, which requires translations to the
supported languages.

The configuration/settings for how required fields must
be presented is passed in as a prop to the public API,
which sets up the necessary configuration context that
can be consumed by components down in the tree.