:bug: [#62] Fix UX issue where invalid dates would clear all inputs
It'd be quite frustrating if a typo leads to other valid input fields
being cleared because the resulting value is not an actual valid
date. So now it's possible to enter invalid dates into the formik
state, which can then be caught by schema validation at the form
level.

This will pose a problem to take into account with form logic - values
are not guaranteed to be valid, though this problem can already present
itself now in the SDK.