:heavy_plus_sign: [#36] Add font-awesome dependency for icons
It's a bit annoying that the icon library leaks down into everything,
but for the time being we're not planning on supporting other icons.

It would be possible to refactor this in the future to make it more
pluggable, so the code to include an icon will be generalized a bit
to make that clear.

It's up to the downstream project to ensure the necessary stylesheets
are loaded - we leave that out of scope because the complexities of
font assets paths are a headache.