:sparkles: [#36] Add an MVP icon library support implementation
For the time being, we stick to font-awesome, but it should possible
in the future to bring in your own SVG-based icons to render or
alternative libraries, so the Icon component has been named in a
generic way to make future expansion more natural.

Plenty of challenges there that we will *not* tackle right now.