:sparkles: [#36] Encapsulate in FieldArray
The editgrid state is now encapsulated in FieldArray, which provides the
helpers to update/replace individual items at a particular index and/or
remove them. It's up to the developer to still properly provide the
array of items, using the same source of form state, but this allows
automatic wrapping in EditGridItem components so that the implementation
details don't have to leak to the outside too much.

This actually unlocks a new feature where we can make per-item decisions
about which controls for saving/removing are available, a thing we need
for the appointment flow with products where you cannot remove the only
product if there's only one.

To enable adding items, only a blueprint/blank item needs to be provided
as a prop (which can also change depending on the state, that's up to
the outer component). Providing `null` for this blueprint disables the
add button.