:sparkles: [#36] Implement the MVP of isolation mode
In isolation mode, changes to an item inside an edit grid are only
applied to the 'main' form state after explicit confirmation/saving. To
achieve this, we set up a separate Formik component with its own context
and state and we invoke the onSave handler to use the FieldArray
replace handler/helper to update the main state.

The edit grid we have also has to support non-isolation mode since we
use this for the appointment products. This is done in a type-safe way
by building on top of discriminated unions based on the isolation mode.

TODO:

* handling validation errors, in particular figuring out what happens if
  we receive validation errors while a nested formik state has been
  initialized already
* passing a validation schema to apply on blur of a form field
* fixing build/type checking errors and integrating it properly in the
  EditGrid main component
* documentation