:technologist: [#36] Use callbacks for item-level content
Specifying callbacks on the root component level
leads to a much better DX because you don't have
to access the raw underlying formik state, which
makes it harder to make mistakes and leads
ultimately to less data wrangling.

We have all the information available about the
form state, so let's let that drive the UI
instead of having to duplicate some information.