:sparkles: [#36] Ensure newly added item is automatically expanded
I wasn't sure if there should be some sort of cleanup function to
reset the state of the indexToAutoExpand when the item has been
rendered initially. It doesn't appear to be necessary, but there could
be an edge case where an item gets unmounted, which resets its own
state and if it then gets re-mounted it could accidentally be set to
initially expanded.

I can't really think of a situation where this would happen, as each
item is rendered by the parent component (EditGrid) for each form
value. Unmounting likely would only happen if isolation mode is changed,
which is a weird thing to happen, or if the EditGrid itself unmounts,
in which case the indexToAutoExpand state would be lost anyway and
the problem goes away.

The only edge case I can see is where you're initially editing in
inline mode, and then it switches to isolation mode. Even then I'm not
sure if it would trigger, since the same component with the same
key gets rendered again, just with different props.