:sparkles: [#36] Display external validation errors in isolation mode
The formik mechanism is to pass errors via the initialErrors prop, which
is suitable for backend errors not set through the client side
validation schema. Or, errors may be set from the validation schema in
the main/parent form and will then be exposed to the nested formik
component by extracting it from the main state.

We look at the errors for the item as a whole in the main context by
composing the lookup path $name.$index, and pass that along to the
isolated item. For inline editing, all of this is already managed
through the main state.

If there are errors, and the item is not expanded, we trigger expansion
automatically so that errors are visible. Especially when these are
errors coming from 'the outside world' they need to be made visible.
When the item is saved (with or without modifications to the data), the
errors for that item are cleared in the parent state so that it can be
properly collapsed again.