:sparkles: [#36] Add label and description to edit grid low level field
For consistency with other field types, we allow specifying the label
and description to the editgrid field as a whole. If no label is
provided, the markup shall not be rendered.

Note that this adds the 'required' asterisk to the label, and if no
label is displayed for a required field, this will not be visible.