:building_construction: [#36] Pass getRegistryEntry callback as prop
This is the best option which doesn't make us do concession on DX or
type safety w/r to the registry. It's still a bit sensitive to
possibly introducing circular imports, unfortunately, but now each
specialised registry entry at least has access to the registry through
it's formField render configuration, which should be sufficient for
higher-level/complex components like fieldsets and editgrids.

This will probably lead to some more refactors/reworks for the
conditional visibility logic when that's being tackled for the
editgrid.

The following approaches were also attempted without success:

* using react context to put the getRegistryEntry in there. This breaks
  the isolation of components (making them have a hard dependency on the
  context), which makes testing in storybook harder as we cannot provide
  these guarantees. Setting up the default value for the context to
  mitigate this introduces the circular import again.
* using dynamic imports to defer loading/populating the registry. This
  should in theory break the circular import chain by moving some
  imports to the runtime, however it didn't work in storybook with HMR
  and requires the entrypoint of the registry to be loaded/hit to ensure
  it loads in type. We also lost type-safety (particularly the
  exhaustive guarantees of the registry) and lost us the type
  information of the modules being loaded. It also break on the target
  build environment since that now includes browsers that don't have
  top-level await, and people with older devices should still be
  properly supported (impoverished citizens, for example).