:safety_vest: [#36] Implement schema validation for edit grid item
Each individual item is now validated using the zod schema created from
the component definition of the edit grid. This validation kicks in on
blur of a field, and on submit of the item form (when clicking the
save button), and it prevents data from being updated when the data
is not valid.

Updated the tests with any to move the 'any' behaviour from the
actual application code to the test, since we know the outcome
of the schema composition are ZodObject instances.