:sparkles: [#78] Allow setting 'outside' errors in the form
Clicking the form field (focus) and the moving focus removes the
error again, which is deliberate - we hook up the validate call as
part of the blur event which runs client-side validation, and these
errors are typically server-side.

That's okay - if the value is still wrong, the next backend call will
re-populate the error(s) again.