:recycle: Update label markup
Following the example on https://nl-design-system.github.io/utrecht/storybook/\?path\=/docs/css_css-form-field--docs,
we no longer render the form field label in an utrecht-paragraph,
but instead we use a plain div with the utrecht-form-field__label
class name.

This fixes the CSS precedence issues for the margin below the label,
which both the paragraph and the form field try to set and the CSS
rules that come last win.

These changes should be visually identical.