:alien: [#93] Add story and fix accessible ID for tooltips
Our custom ID was being overwritten by the tooltip hook props, which
uses useId internally to relate the tooltip trigger and tooltip
content to each other. The initial expectation was that this was wrong
behaviour and we want to have the tooltip referenced too in the textbox
aria-describedby, but after testing with Robin's macbook voiceover, the
tooltip itself is properly announced and including it twice is odd. So,
we've decided on *not* referring to the tooltip from the textbox (and
other form elements) itself.

Doing otherwise requires some patching of the floating-ui context too,
which feels not very kosher, so perhaps it's another hint that my
initial approach was wrong.