:sparkles:[#59] Implement editgrid-item scoped simple conditionals
The scope of the references to the data is now properly maintained
in the simple conditionals, i.e. a condition trigger field key is
looked up in the main/parent scope, while the item scope is managed
by overriding the value of the whole edit grid to point to a
particular path. This makes looking up other items by index not
possible I think, but I'm not sure if Formio even implements this
or whether this is an actually used feature.

I have some doubts about nested editgrids still though, need to
add a test to check the correct behaviour there.