:sparkles: Support simple conditionals in (nested) editgrids
Formio pulls some tricks to make simple conditionals possible within
the scope of a single item while still allowing conditions to refer
to components outside of the edit grid.

Where normally the editgrid component key has an array type as value
in the submission values object, within an editgrid item context they
actually replace this - the values of the item itself are placed under
the key of the editgrid it the submission data object. This results in
a merge of the root values together with 'editgridKey.componentKey'.

Simple conditionals that refer to a component inside the edit grid have
this prefixed reference for that reason.

Our formio renderer now replicates this behaviour, making it possible
to make a component inside an editgrid conditionally visible/hidden,
and the reference component can exist both outside of the editgrid or
inside of it. This is also taken into account in the item preview mode
when it's not being edited.