:safety_vest: [#106] Run selectboxes validation on blur
This turned out to be surprisingly hard to get right because of the
order of DOM events firing. The desired behaviour is:

* do not validate when the checkbox receives/has focus, as the user
  is still in the process of filling out the field
* don't validate when the move from one checkbox to another, as they're
  still in the same field/component
* run validation as soon as the field/component as a whole loses focus,
  so that feedback is provided as soon as possible without being a
  nuisance

What happens in browser/DOM terms is that clicking a checkbox triggers
a focus event on the checkbox, then clicking another checkbox first
triggers a blur event followed by a focus event on another element.
However, for a brief moment, document.activeElement points to nothing
(except the document/body itself) while transitioning the focus. If
you click the input directly, then the blur event has a relatedTarget
indicating the next element that will have focus, but if you click the
label this relatedTarget is unset, as the label click itself triggers
another click event on the input.

The above basically makes it impossible to subscribe to onFocus or
onFocusCapture or even low-level focusout/focusin events, as there's
always a moment that the focus is outside of the fieldset even when
it will go to another checkbox inside -> this is a race condition.
Attempted workarounds were: setTimeout 0 to schedule a new macroTask,
but this still fires before the new focus happens by the browser, and
the same applies for requestAnimationFrame. Using a 10ms timeout was
still too early, while it did work reliably with 100ms timeouts, but
MagicNumbersBad + this produces lag for the user.

Instead, I've opted to listen to focus and click events on the whole
document:

* a focus on any element outside of our fieldset/container/checkbox
  group is guaranteed to have caused a blur on our component, so we
  can safely trigger validation
* a focus on any element inside the container can be safely ignored,
  since at the very last form submission will result in a blur and it
  also triggers validation by itself. Plus, navigating to any other
  element will result in a focus event on another element, which brings
  us back to the previous situation
* clicking anywhere but labels/inputs triggers validation as those
  cause blur events. We single out labels and blurs because those
  result in a focus event (anywhere on the page) and are handled by
  another event handler already.

Note that a label click event also triggers a click event on the
related input, which is not relevant for us, but it may be
surprising.