:package: Fix packaging mistake
The vitest.setup.ts inclusion in the TS config was changing the
root from src to the git root, causing the declaration files
to not be emitted anymore.

We don't want to distribute this file anyway, so in the
production config it's included, fixing the generated
type declarations in the package.