:alembic: Experiment for single-pass clearOnHide
This re-evaluates the component visibility but only determines and
updates the result of clearOnHide once. The resulting values are fed
into the visibility check again (so the visibility check if effectively
double pass, while clearOnHide is single pass).

I think I'd rather keep this 100% reactive and let the values/visibility
state converge, but there are some problems with this approach:

* the behaviour is different from Formio.js SDK itself
* we would need a good cycle detection to prevent infinite rendering
  loops, which single-pass does not suffer from

This, however, is some complicated code and we're not there yet.