:fire: [#59] Delete failing test
The expected behaviour expressed in this test applied only when
clearOnHide behaviour was not yet implemented. However, due to the
expected behaviour when a component becomes visible again from being
hidden *and* an implementation strategy of being reactive, the
original behaviour is no longer achievable.

We don't see this as a problem though - it makes little sense
semantically to remove submission data values for components that are
displayed on the screen - this is not intuitive. Additionally, there is
a practical limitation that the backend cannot possibly serialize
'undefined' for Javascript to apply the data changes because of
backend logic rules, and that feature is the main reason why we need
these imperative handles in the first place.