:sparkles: [#108] Add components data to the FormSettingsProvider
The SoftRequiredErrors component needs information about all components in the form. The best way to do this, is to provide `components` data in a context provider.

As `components` data is a piece of form global information, we can add it to the FormSettingsProvider. Also, when the `components` data changes, we want to/already re-render the entire form anyway (so by adding `components` to the context provider we don't introduce more re-rendering).