:building_construction: [#59] Rework how we handle component visibility
The previous iteration *did* take into account form/submission values
to evaluate conditional visibility, but didn't account for the side
effects this produces through the `clearOnHide` feature.

The short summary is: values determine whether a component is visible
or not (so they affect 'component configuration'), BUT this is not a
one-way street - the visibility state/configuration of a component
affects values through the `clearOnHide` behaviour.

The code dealing with this has been renamed and restructured. The main
idea is:

* as input, take an array of components of which you want to see the
  visibility state, and an object with values for the conditional
  evaluatio scope
* as output, you get an array of visible components and updated values

The component array recurses so that layout components can apply the
visibility state to their nested children immediately. Values are also
updated immediately, and the updated result is fed to the next
component and/or item, which helps minimize evaluation cycles.

The result is put back in the React component/Formik state, i.e. the
values are applied back to the formik state *if* there are updates. We
are very careful in only mutating when needed, using the setIn and
replace helpers from Formik. If the input values are equal to the
output values, we guarantee that the object identities are the same
and React will not render unnecessarily (otherwise you'd end up in an
infinite render loop).

This architecture also deliberately separates the 'submission values'
and 'evaluation scope' concepts. Most of the time they're the same,
but Formio's edit grids work differently, where each item is
essentially assigned to a different scope when evaluating it. This
evaluation scope can not be assigned back to the item values, since it
contains values from its outer scope too. The extraEvaluationScope that
existed before has been replaced with a callback so that the parent
caller can provide the scope through a closure, as only the parent node
has the necessary context.