:sparkles: [#59] Ensure that edit grids support clearOnHide behaviour
The architecture rework was needed because of edit grids because the
simple model of layout components doesn't apply here:

* the nested components definition is not static for each item
* there is no explicitly defined components configuration for each
  item, instead they're 'virtual' and the blueprint is provided by
  the edit grid component itself
* when dealing with nested edit grids and their preview/edit mode, you
  can't rely on the React/Formik state to persist the updated values
  back into the parent value/state

So, this rework splits the value/state management and the presentation
logic. `applyVisibility` is implemented and called recursively from
the top-level component. It's responsible for making sure the Formik
values state is synchronized with the visibility state. It can't assign
a modified list of components for each item anywhere. It is active at
all times, whether an item is in preview mode or edit mode.

The `ItemBody` and `ItemPreview` components also invoke the
visibility processing. The former synchronizes the updated item values
back to the Formik state, while the latter is purely presentational. It
must be present, because for nested edit grids, ItemBody is not called
for preview mode.

This setup does simplify the item preview, because we now have the
guarantee that the proper evaluation scope is passed to `ItemPreview`.