:white_check_mark: [#59] Add tests for cycle detection in edit grid
Edit grids have access to the outer scope, but typically you don't or
can't specify the outer scope to depend on a (particular) component
inside the edit grid, so no cycles should be detected there.

Note that the above *may* be more nuanced if you can target the
edit grid as a whole, but we need to check if that's possible in
vanilla formio in the first place.

However, within the components of an edit grid you can definitely
have cycles, and those need to be detected or Bad Things will happen.