:building_construction: [#59] Defer conditional value test to component registry
Instead of performing the value check centrally where knowlegde of each
shape of value and associated component type is required (which isn't
even possible in the first place), check if the component in the
registry provides a comparison check/callback to perform the check. If
none is provided, we fall back to a default and simple/naive
implementation that covers the majority of cases.

This allows us to properly interpret the component value in the right
context instead of guessing or assuming what the datatype of the
value means - which is what formio.js does.