:bug: [#133] Fix infinite render because of wrong value mutations
The scope calculation in nested edit grids was not correct. Rather
than deeply setting the deeply nested item values in the evaluation
scope, every level was being added to the same root depth of the
scope.

This caused two problems:

* the submission data was wrong, because the hidden values are not
  properly being cleared and updated in their correct scope
* it leads to mutations in the wrong place in values, which leads to
  unintended parent value mutations which serve as input again for
  the next render loop, thereby causing infinite renders

Fixing the data updates/scope calculations solved both issues.

The mechanism to work with nested scopes was wrong, now it works on an
inside-out principle: the most inner scope is set, and that value is
passed to the parent scope calculator which sets its own level etc.
all the way back up the root node. This way each layer 'scopes' the
values with the appropriate parent key.