:bug: [#141] Proper handling of edit-grid hidden
The infinite re-rendering of components when an edit-grid is hidden, came from the edit-grid values set by `applyVisibility`. By making sure that the values reflect the actual state, we can prevent infinite recalculations and re-renders.

When the edit-grid became hidden, its value would reset to undefined. The `applyVisibility` would then make sure that the value for editgrid was at least an empty array, which was different than the value calculated by processVisibility, which would cause a new calculation/render, which replay this proces infinitely.

By making sure that empty edit-grids are kept as they are, we cannot have data mismatches, so we won't have re-renders.

because it takes 1 render cycle for the edit-grid values to be re-replied, we have to handle a value of `undefined` in the EditGrid component.