:sparkles: [#97] Add legacy cosign component
Ideally we would like to keep the folder name the same as the component type, but it differs from
the cosign v2 component only by casing, and typescript complains about it (TS1149). Also, as far as I
understand it, there are differences across operating systems whether paths are case-sensitive.
Therefore, it seems like the safest option to name the folder `cosignOld`, instead of `coSign`.