:sparkles: [#119] Add Loader component
Added a simple Loader component, which shows a basic spinner and contains accessible text for screen-readers.

To properly alert accessibility tooling about the loading status, the Loader component must always be rendered. This is because the container, which has the role `status`, most be present in the DOM before the content changes. See https://www.w3.org/WAI/WCAG22/Techniques/aria/ARIA22 for more information