:sparkles: [#105] Implement react-select wrapper
Added a wrapper around our use cases for react-select, taking into
account:

* styling with NL DS (build on top of utrecht-select)
* translate the form state values from Formik into the selected
  option(s) passed to react-select

This is *mostly* a port of the Select in the SDK, polishing some
typescript bits here and there and iterating on the isMulti variant
with regards to styling.

A future pass to properly set up the class names/design tokens will
still be necessary.