:wheelchair: [#105] Ensure that the select input is described by the error message too
The aria-describedby is not supported out of the box in react-select,
but there is a workaround where we can grab the select props and check
any additional aria-describedby IDs to add.

By default, the placeholder text for the search input is used as
accessible description for the input.

As usual, aria-errormessage cannot be used yet because of poor
screenreader support, which is a shame because this prop *is* supported
by react-select.

See https://github.com/JedWatson/react-select/issues/1570 for the
upstream issue and the inspiration for the workaround.