:alien: Handle broken conditional specifications on component
Not specifying a conditional in the builder results in formio.js adding
a conditional which has 'null' values for the show and when
properties, which are *not* included in their own 'ConditionalOptions'
type.

Strict equality checks on undefined therefore result in us incorrectly
thinking that there is a conditional specified, which overrides the
static 'hidden' property and breaks expectations.

This patch uses non-strict equality which treats undefined and null
the same way, working around the issue.

The patch can be reverted once we don't use formio.js anymore in the
builder *and* once the existing form definitions in the database in
the backend have been migrated.