:bulb: Add a TODO to the label wrapper
The browser dev tools warn about using <label> within a <legend> that's
not linked to any form input, which breaks the semantic markup. We
currently use the Utrecht label component for the styling, but this
should (for radio/selectboxes) instead be a plain span instead of
label tag, as the legend itself already captures the semantic
meaning.