:building_construction: [#111] Add component-dependency injection mechanism
Some components (like addressNL but also the file upload component)
require some additional tooling/dependencies to perform part of their
function, most commonly backend API interaction.

Since we want to keep the formio-renderer package pure and free from
API client details, we set up a dependency injection mechanism via
the existing form renderer settings context. The caller (our SDK in
this case) can then pass in the necessary configuration/callbacks
to perform the API interaction, as long as it conforms to the
specified interface.

This adds an interface for the addressNL component where given a
postcode and house number as input, it gets back a resolved city +
street name (or empty values if it could not resolve), or `null` if
there was some kind of unforeseen error.