:heavy_plus_sign: [#111] Add react-use dependency
Added react-use to be able to use the useAsync hook for the async
postcode lookup (injected as a dependency).

It was briefly considered to not use a dependency for this, but looking
at the implementation it handles a lot of cases like the component
being mounted and still being in the right render cycle to apply the
result, that I didn't feel confident in 'rolling our own'.

Added as a peer dependency with a version range compatible with our
current SDK so that we can leverage tree shaking optimally and optimize
bundle sizes.