:sparkles: [#111] Implement obtaining initial Formik values
A well-defined interface is not assignable to JSONObject because it
doesn't define the string key access, which is a bit annoying. For now
we stick to a cast, because in spirit it is a JSON-serializable type,
but we have some other places too where more detailed information
proves challenging with generic types so we may revisit this once 1.0
is available.