:sparkles: [#111] Implement the address NL value display (for editgrid)
The 'string representation' of an address.

Currently there are two cases - address derivation enabled or disabled.
In the latter case, we don't have a city and/or street name, so we can
only display the postcode + house number (+ extra parts), in the former
case we have extra information and can format it according to the post
system formatting guidelines.

Note that the official recommendation is to display the postcode with
a space between numbers and letters, and the separation between
postcode and city should be two spaces. There are no official
guidelines for the separator between house letter/number and house
number addition, but a dash or forward slash are common.